Remove OnRuntimeUpgrade hooks (#614)
* Remove OnRuntimeUpgrade hooks


* [no ci]Update changelog


* [no ci]Fix changelog


* Move storage migration to migration file


* Fix build
